(packaging) Use the packaging loader for tasks
The packaging repo now uses an explicit loader which handles loading the
various rake tasks in packaging. This commit updates the Razor Rakefile to
use the loader instead of a glob of the ext/packaging/tasks directory. We move
this load into the rescue LoadError block because the packaging repo won't
always be there, but if it is, we want to load it.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>