Fix for trusty's new sshd_config.
On trusty onwards PermitRootLogin is set to without-password which
means Razor can't log in to set up the broker stuff. This fixes that.

I'm using the precise model for everything because apart from this
change (which is a no-op on everything else), it just works. And it
needs replacing with the new Razor anyway.