Simple local hack to modify adduser behaviour.
This installs a customised adduser.conf (changed system UID/GID
ranges) in to the target partition before the operating system gets
installed. This ensures that all system accounts created during
installation have the correct UIDs.

This is Kent specific, and the range chosen matches our policies.

This could also be cleaner and more generic, but since this version of
Razor is deprecated I just went for the minimal effort hack.