Force the grub device for multi-disk systems.
I'm not sure if this is due to a change in a newer trusty installer,
but recently in multi-disk systems it's been prompting for which
device to install grub on. In this case it'll always be /dev/sda, so
it's fine to hard code it.

Not sure if the same needs to be done for the RAID one.