Always install the latest Puppet version.
This adds the PuppetLabs APT repository before installing Puppet
so that we always have the latest version.

Note that we still install the OS-provided facter during the OS
installation because it's needed by os_boot.erb to work out the
node's IP address. We could probably find a better way to do that,
but it doesn't matter too much. I've opted to just remove the
OS-provided facter and dependencies before installing the PuppetLabs
ones.