Add required GPT boot partition.
This is only used when GPT is used, which at the moment only happens
with big disks. There's other problems with that (/boot can't be
past the 2TB point on the disk) which this doesn't fix though. It's
in the default recipes, so worth including here for future use if
GPT becomes the default.